pyproject.toml
[project]
name = "aisearch"
version = "1.0.0"
description = "Tavily compatible search service based on SearXNG and Crawl4AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "beautifulsoup4>=4.13.3",
    "crawl4ai>=0.5.0.post8",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "markdown>=3.8",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.0",
]
